openjdk-8: fix TUNE_CCARGS on intel-core2-32/i686
A machine-independent recipe must not have machine-specific changes,
as in TUNE_CCARGS_append_intel-core2-32. In the best case, that only
causes unnecessary rebuilds when switching machines with the same CPU
which otherwise could share sstate. In the worst case, it confuses PR
server and build history such that building fails with "version going
backwards".

There are not tune-specific overrides, so a Python snippet is used to
apply the additional parameters on all machines with the same
TUNE_ARCH as intel-core2-32 (like for example edison).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>
Signed-off-by: Brendan Le Foll <brendan.le.foll@intel.com>