openjdk-8: Fix openjdk build when CXX contains assembler arguments
Builds currently fail, if CXX also contains assembler arguments. These
arguments are in form "-Wa,args,are,here". When this string is then used in
openjdk make process, it is inserted before the make evaluates its argument
lists. As these lists are comma separated, the commas in assembler argument list
are interpreted as list separators, which breaks the build.

This is fixed by adding the extra '$', which means the string is inserted after
make evaluates the list.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>