openjdk-8: Patch compiler flags in adlc build file
ADLC is a tool that is used when compiling hotspot C2 compiler. As it is used
during build phase, it is a native tool that uses native toolchain. It however
uses compiler flags through CXXFLAGS and CFLAGS, which have values set for
cross compilation by bitbake (CXXFLAGS=TARGET_CXXFLAGS. CFLAGS=TARGET_CFLAGS).

This causes issues with older compilers, as currently the target flags have
the -fstack-protector-strong set through security flags. This flag is only
supported by GCC 4.9+, which means compiling openjdk 8 will fail, for
example, on Ubuntu 14.04 LTS.

This is fixed by replacing the target flags with build flags, which are
intended for the build host.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>